+++ packages/bloc_test/test/blocs/blocs.dart
@@ -3,6 +3,7 @@ export 'complex_bloc.dart';
 export 'counter_bloc.dart';
 export 'debounce_counter_bloc.dart';
 export 'exception_counter_bloc.dart';
+export 'instant_emit_bloc.dart';
 export 'multi_counter_bloc.dart';
 export 'side_effect_counter_bloc.dart';
 export 'sum_bloc.dart';