+++ packages/bloc_test/test/blocs/side_effect_counter_bloc.dart
@@ -2,23 +2,21 @@ import 'dart:async';
 
 import 'package:bloc/bloc.dart';
 
-enum SideEffectCounterEvent { increment }
+import 'blocs.dart';
 
 class Repository {
   void sideEffect() {}
 }
 
-class SideEffectCounterBloc extends Bloc<SideEffectCounterEvent, int> {
+class SideEffectCounterBloc extends Bloc<CounterEvent, int> {
   SideEffectCounterBloc(this.repository) : super(0);
 
   final Repository repository;
 
   @override
-  Stream<int> mapEventToState(
-    SideEffectCounterEvent event,
-  ) async* {
+  Stream<int> mapEventToState(CounterEvent event) async* {
     switch (event) {
-      case SideEffectCounterEvent.increment:
+      case CounterEvent.increment:
         repository.sideEffect();
         yield state + 1;
         break;