+++ packages/bloc_test/test/cubits/cubits.dart
@@ -2,6 +2,7 @@ export 'async_counter_cubit.dart';
 export 'complex_cubit.dart';
 export 'counter_cubit.dart';
 export 'delayed_counter_cubit.dart';
+export 'instant_emit_cubit.dart';
 export 'multi_counter_cubit.dart';
 export 'side_effect_counter_cubit.dart';
 export 'sum_cubit.dart';